@@ -4,14 +4,14 @@ version: 1.3.2
 homepage: https://github.com/and2long/flutter_ytim
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
   web_socket_channel: ^2.0.0
-  dio: ^3.0.10
+  dio: ^4.0.0-beta5
   provider: ^5.0.0
   shared_preferences: ^2.0.3
   crypto: ^3.0.0
