@@ -3,7 +3,7 @@ import 'package:flutter_ytim/src/bean/im_msg.dart';
 
 /// 全局IM信息
 class IMLastInfo {
-  IMMessage msg;
+  IMMessage? msg;
   int unreadCount;
 
   IMLastInfo({this.msg, this.unreadCount = 0});
@@ -11,13 +11,13 @@ class IMLastInfo {
 
 /// 全局IM信息，支持刷新
 class IMStore with ChangeNotifier {
-  Map<String, IMLastInfo> _lastInfos;
+  Map<String?, IMLastInfo> _lastInfos;
 
   IMStore(this._lastInfos);
 
-  Map<String, IMLastInfo> get lastInfos => this._lastInfos;
+  Map<String?, IMLastInfo> get lastInfos => this._lastInfos;
 
-  void update(Map<String, IMLastInfo> map) {
+  void update(Map<String?, IMLastInfo> map) {
     _lastInfos = map;
     notifyListeners();
   }
